requirements.txt
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.14.0
yfinance==0.2.65
pandas==2.3.1
google-genai==1.31.0
gunicorn==23.0.0
numpy==1.24.3
scikit-learn==1.3.2
openai==1.12.0
pydantic==2.5.0
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99